employee_app/requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.0.0
pydantic>=2.7.0
email-validator>=2.0.0
pytest>=8.4.0
pytest-html
Flask-CORS>=4.0.0
PyJWT>=2.8.0
Werkzeug>=3.0.0
